[DEV-5790] Define web models for search prs filters

+++ server/athenian/api/models/web/goal.py
@@ -16,7 +16,9 @@ class AlignGoalsRequest(Model):
     include_series: bool = False
 
 
-MetricValue = float | int | str
+# the order is from more specific to less specific
+# since models deserialization tries the types in order
+MetricValue = int | float | str
 
 
 class GoalSeriesGranularity(Enum):
